Moved entities_dict.py to metadata folder
