+++ bam_masterdata/metadata/entities_dict.py
